Disabling Recovery Reverts Finalization (#27)
Closes #5

This PR adds a rule which checks that the finalization of the recovery
process is not possible for a safe account that has disabled the
recovery module.

The prover found a counter-example which allows calling the finalize
function with zero new owners and threshold from a safe having exactly
one owner. Recovery initiation with zero new owners is not possible,
thus this counter-example is not considered (but still we check the
owner structure is not changed for the safe account after the finalize
function call).