Inherit guardian storage (#34)
This PR changes the design of how the guardians are stored.

Prior to this change, the guardians were stored in a separate contract
`GuardianStorage`. The `GuardianStorage` address was then passed as a
constructor parameter to `SocialRecoveryModule` which could then make
changes to the list on request by the user. So, the recovery
functionality was achieved by having two contracts and thus, two
storages.

As a side effect, it was initially brought to light that any other
module can also modify the list of Guardians in `GuardianStorage` and
was not write-restricted only to `SocialRecoveryModule`. Any other
enabled module on a Safe could change this list.
See:
https://github.com/5afe/CandideWalletContracts/blob/870f82d3c4d286b3fd9241c6479c9472d7d73fc6/contracts/modules/social_recovery/storage/GuardianStorage.sol#L35

The initial argument for accepting this behavior was that any malicious
module can:
1- Change ownership of the account
2- Steal all account's funds
3- Execute any arbitrary function of the `SocialRecoveryModule`
impersonating the owner
Thus, it did not violate any security assumptions.

Also, allowing the guardian list to be edited by any module allows
upgradability of `SocialRecoveryModule` without having to migrate the
list of guardians.

But, later auditing partner also reported the same behavior as an issue
with the argument:

- There should be a distinction between malicious modules (or any
modules that can perform arbitrary operations) and modules that are
limited to do these things by their implementation. These "well
behaving" modules will effectively gain more privileges than they
actually need.
- It opens a new attack vector and violates least privilege principle.

After internal discussion, we concluded that allowing any module to
modify the guardians list should not be permitted and this PR reflects
the same.

In this PR, `SocialRecoveryModule` inherits `GuardianStorage` so that
there is only 1 contract involved in recovery functionality. By making
this change and removing `onlyModule` modifier, and other changes in the
code (removing `_wallet` as a parameter in some functions) now allows
only wallet i.e. `msg.sender` to change the guardians.

This also simplifies the design of the recovery functionality and
enforces to use storage space of a single contract i.e.
`SocialRecoveryModule`.

Other changes in PR include:
- Fixes to files related to formal verification
- Removing duplicated code which is now not needed
- Remove redundant parameter `_wallet` in some functions

---------

Co-authored-by: Shebin John <admin@remedcu.com>
Co-authored-by: Nicholas Rodrigues Lordello <n@lordello.net>