Invalidate `nonce` (#40)
This PR adds an extra function called `invalidateNonce(...)` based on
the recommendation from the audit team to invalidate a particular
recovery process. It could be used in two ways:
- Assuming the owner lost access to the Safe, the owner requests a
guardian for recovery initiation. Once the recovery initiation started,
the owner realized that the access was not actually lost, thus to
invalidate the current initiation (not execution, for that owner can use
`cancelRecovery(...)`), the owner can actually call the
`invalidateNonce(...)`.
- As a best practice to invalidate any `confirmedHashes[...][...]` from
a guardian which is soon to be removed.

Also added:
- a test to check if execution fails in the case of an invalidated
nonce.
- run test in CI.
- FV to check if invalidating a nonce cancels an execution of recovery.

---------

Co-authored-by: Nicholas Rodrigues Lordello <n@lordello.net>