Dispatcher based update (#45)
This PR updates the summary with `DISPATCHER(false)` in
`RecoveryConfirmationSignatureValidity.spec` as for most cases, it works
as intended (i.e. find the correct implementation from the available
contracts) and a summary is not required (to include possible reverting
paths as well).

Also, we found one case which should be checked (`isModuleEnabled(...)`)
but was ignored because we used a summary before.