Dispatcher update in `SocialRecoveryModule.spec` (#46)
This PR updates the summary to use `DISPATCHER(true)` for the
`execTransactionFromModule`.

From Certora Team:

> DISPATCHER(false) and DISPATCHER(true) both first try any known
Contract. The difference should be only if a different contract is
called which is not explicitly linked into the scene. With
DISPATCHER(true) the prover just assumes that this will never happens.
It only tries those input values for which you only call known
contracts. With DISPATCHER(false), the prover will also consider the
case you call an unknown contract and in that case assumes the worst,
i.e. all ghost variables are havoced and the return value is arbitrary.
So you get a lot of spurious counterexamples with DISPATCHER(false), but
with DISPATCHER(true) you are technically unsound, as you don’t verify
anything if unknown contracts could be called.

So, In our case, as we already summarized call to
`execTransactionFromModule(...)` if `callee` is `safeContract` before.
And as the summary goes for the non-reverting path, I think it is safe
to use `DISPATCHER(true)` here.

P.S. For the `RecoveryConfirmationSignatureValidity.spec`, as we didn't
require some method specified, that was removed, along with an extra
harness function.